Remove Emojis
Hey Developer. This change us exclusive to uYouEnhanced since a lot of them at that Repo doesn’t seem to like the emojis. Maybe there can be a reconsider about changing it back? Or possibly turning it into an option? That could work but not sure if it’s possible. But anyways thanks again for your tweaks PoomSmart!